requirements.txt
attrs==17.4.0
boto==2.48.0
boto3==1.7.10
botocore==1.10.10
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
Django==2.0.4
docutils==0.14
funcy==1.10.1
future==0.16.0
gensim==3.4.0
idna==2.6
image==1.5.20
Jinja2==2.10
jmespath==0.9.3
joblib==0.11
kiwisolver==1.0.1
lda==1.0.5
MarkupSafe==1.0
matplotlib==2.2.2
more-itertools==4.1.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nltk==3.2.5
numexpr==2.6.4
numpy==1.14.2
pandas==0.22.0
pathlib==1.0.1
pbr==4.0.2
Pillow==5.1.0
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
py==1.5.3
pyLDAvis==2.1.1
pyparsing==2.2.0
pytest==3.5.1
python-dateutil==2.7.2
pytz==2018.4
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
smart-open==1.5.7
spacy==2.0.11
stop-words==2015.2.23.1
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.1
ujson==1.35
urllib3==1.22
wordcloud==1.4.1
wrapt==1.10.11
